model/requirements.txt
# Requirements for Graph WaveNet Model

# Deep Learning
tensorflow>=2.14.0
keras>=3.0.0

# Graph Neural Networks
spektral>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# BigQuery
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0.0
